WIP refactor sessions queries
Todo
- Fix test_sessions_list test
- Move tests to posthog.queries
