+++ posthog/api/test/test_event.py
@@ -247,6 +247,9 @@ class TestEvents(TransactionBaseTest):
 
         with freeze_time("2012-01-15T04:01:34.000Z"):
             response = self.client.get("/api/event/sessions/").json()
+        import ipdb
+
+        ipdb.set_trace()
         self.assertEqual(len(response["result"]), 2)
         self.assertEqual(response["result"][0]["global_session_id"], 1)
